BACKPORT: sf: hwc1: Set SF EventThread to SCHED_FIFO in all cases
Previously, we were only setting the SF EventThread to SCHED_FIFO
when the phase offsets aren't the same. This would probably result
in more jitter for devices whose phase offsets are the same.

This was the result of the Android 7.1 mismerge.

This essentially reverts commit 12f5891
and fixes the Android 7.1 mismerge.

Change-Id: I73594001e381c176410e06865352dff74019ff27